+++ docs/.vitepress/theme/index.js
@@ -1,9 +1,0 @@
-// .vitepress/theme/index.js
-import DefaultTheme from 'vitepress/theme'
-import './custom.css'
-import Layout from './Layout.vue'
-
-export default {
-  extends: DefaultTheme,
-  Layout
-}